requirements.txt
Flask==3.1.3
gunicorn==22.0.0
Werkzeug==3.1.8
psycopg2-binary==2.9.9
pandas==2.2.2
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
python-dotenv==1.0.1
Jinja2==3.1.6
